Skip already imported resources
Add a `Imported` flag for each import item, if it is imported then set
that flag. In case users hit some import error, the users can adjust the
resource types for each import item. In case users adjust the resource
type for an already imported resource, the tool will remove that
resource from the state and mark it as not imported, which will cause
this resource to be imported in the next apply.