Optional overwrite outputdir (#31)
* Refactor meta initialization

Lift up meta construction before the UI program starts. The meta
construction now is only responsible to convert the cfg to meta/model,
do some simple check (e.g. whether the output dir is empty), while move
the time consuming stuff like downloading `terraform` to the
meta inialization phase.

The newly added `-f` option is just a placeholder here.

* Option to allow overwrite output dir

The `-f` option can be set to overwrite the outdir if non-empty.
Otherwise if `-f` is not set, in interactive mode, the users are prompt
to choose whether to overwrite the dir if non empty.