Refactor: remove dependency on the v1.1.0-alpha terraform
Remove the dependency on the v1.1.0-alpha `terraform` only for the `terraform add` command. Instead, depends on the github.com/magodo/tfadd.

Correspondingly, removing the SDK schema generator and generated files, which is offloaded to the github.com/magodo/tfadd. But we still need to generate the resource mapping here.

Additionally, as the `terraform-exec` now depends on the `hcl-install`, we are also migrating to using it. However, instead of using the upstream one, we use a forked repo, which has an additional feature to specify version constraint when searching for any fs source.