Special Id transform + special post HCL lifecycle addon
Now the resource listing logic is as below:

- ARM template export into its raw format
- Populate managed resources on top of the ARM template
- Convert the ARM template raw format to fully qualified format with
  subscription id and resource group name. Especially, the resource id
  (both as the resource identifier and as the dependency) is converted
  from its ARM resource id expression form, to its corresponding TF
  resource id literal (this might requires some special transformation
  based on the different formats used between ARM and TF)

From this point on, all the processing is on top of the FQ template,
with only the TF resource id taken into consideration. This makes a
change on the UI, where it previously showed the resource id in ARM
format, whilst now it shows the TF resource ID. This also means the
saved and imported resource mapping file is now using the TF resource
ID format.

After resources are imported and config is generated. Now we not only
adds on the dependency on top of the generated HCL, we also add on for
some necessary lifecycle meta arguments, `ignore_changes` only for now.
This is necessary for resources that by design the service will create
some "known after apply" properties, which normally users will want to
ignore, so that the follow up plan will not introduce diff. An example
is the `tags` for the application insights webtest resource (see the new
e2e test added for details).

Some follow-ups:

- The ID transformation hook can be extended to transform the ID from
  its mgmt plane ID to its data plane alternatives (e.g. for key vault
  nested items)
- The ID transformation hook should introduce a cache, to avoid
  same expensive conversions from running multiple times