Conslidate the user resource mapping file
Previously, the saved resource mapping file only contains the items that have specified TF resource type. The skipped ones are absent. This is not friendly to resource group that has a large scale of resources, and the current auto deduction capability only provides users a limited set of resource types. This results into a lot of resources missing in the generated resource mapping file.

This PR changes this situation that all the resource listed via ARM template will be in the generated resource mapping file. The skipped ones will be left in the json object with empty string value. To simplify things, we regard empty `TFAddr.Type` as to be skipped in the import list. Especially, when importing with resource mapping file, we will not do auto deduction for resources, which means only the import item specified in the resource mapping with a non empty TF resource type will be marked to import (though users are still able to specify the others to import during the interactive session).

Besides, this PR fixes a UI bug for progress view that is introduced in another PR.