Do once empty check on the outdir prior to TF run (#180)
TF run with `TF_LOG_PATH` set to a local file will cause the empty check
on the output dir return non-empty on the second check in the `Init()`
(actually `initProvider()`) function. In this case, the existing logic
will work in append mode, especially it creates the `provider.tf` with
no `terraform` setting, i.e. no provider setting.  This results into
inconsistent dependency lock issues when users are not opt in the
`dev_overrides`.

Meanwhile, this commit explicitly set the `TF_LOG` and `TF_LOG_PATH`
from the environment, for the terraform-exec library. So that users can
inspect what is going on via Terraform.