Resource mapping file is matched case insensitvely, and the resource id is used for import
When providing the resource mapping file, it is matched case sensitvely before. Changing it to insensitive is to mitigate a bug that ARG will return its first resource id with RG name uppercased, which will cause the mismatch if compares case sensitively. Additionally, the resource id that is recorded in the mapping file will be used as the tf id for importing (case sensitively). This is useful when aztft has bugs that returns the incorrect casing for certain resources, users can tune the resource mapping file to get the correct casing. Whilst there might be cases that the resource id is not the same as tf id not only in casing. In that case, we will need to further modify the format of resource mapping file to accomodate it.