Discover reference dependencies based on HCL
There is another option that discovers the dependencies on API models, just like ARM does. This might save us the dependency for cases like VM disk depends on VM, which is now missing because that dependency is introduced by the attachment resource. However, that can only save a very small amount of cases, but there are still a big part of dependencies can't be discovered by simply looking into API properties, which are heavily depending on the business (e.g. sentinel resources depends on an operation insight resource).

For catching up the gap of the previous VM depends on disk, we shall consider introducing the association resources in the azure resource set hack code. This is the great place for holding such kind of code, as we are already using it to populate the missing managed disk in this case. If we go through this direction, then discovering reference dependencies based on HCL is the right choice here, as it will then be able to add on the dependencies for the association resource on the associated resources.