Resourceset hack populates `azurerm_virtual_machine_data_disk_attachment`
This commit introduce a new field `PesudoResourceInfo` in the `AzureResource` to record TF pesudo resources, which will later be used when converting into `TFResource`. In this case, rather than using aztft to look for the TF resource type based on the concrete azure resource, it directly consume the info recorded in the `PesudoResourceInfo`.

With above, now resourceset hack not only populates the managed disk, but also the disk attachement resource.