Handle multiple azure resources have the same TF id when resolving TF dependency (#251)
There are cases that multiple Azure resources correspond to the same TF resource id. E.g. parent and child resources, or an association resource and its master resource. In this case, when resolving cross resource dependencies, as we are depending on the TF resource id, we will ends up with multiple TF resources. There is no good way to tell which one of them is the exact depended resource. The only right thing to do is to add a comment containing all the candidates and let users to choose.