Avoid `terraform init` for import workspaces when `--dev-provider` is specified
For environment that has `dev_overrides` correctly setup for the providers, users can use `--dev-provider` to indicate instead of using the pinned version, use the local dev provider. Another effect of using `dev_overrides` is that it will not check the dependency lock for each terraform command, hence avoid calculating the provider hash everytime. This saves us a bit of CPU time, especially we are running `terraform` commands in parallel.

Therefore, this PR extends the `--dev-provider` a bit to not only remove the version constraint in the `terraform.tf`, it will also avoid the `terraform init` for the import directories, as the follow up terraform command will not use that file at all (given `dev_overrides`). This saves us some CPU time during `Init()`.