Integrate with terraform-client-go for hcl only mode to eliminate `terraform` binary (#381)
* Integrate with terraform-client-go for hcl only mode to eliminate `terraform` binary

Integrate with terraform-client-go for hcl only mode to eliminate `terraform` binary. This improves the performance (quicker speed and lower footprint) of importing.

To use it, users have to explicitly specify a new hidden flag (`--tfclient-plugin-go`). This flag is hidden since it is not supposed for CLI users to use, but is meant to be used by module users who wants to get the benefit of performance improvement.