Support implicit dependency based on top level `resource_group_name` attribute.
The dependency injection logic now also add supports for replacing the top level `resource_group_name`'s value by a reference to the resource's `name` attribute if that resource's name has the same value, and that resource is a resource group. This is almost only relevant to `azurerm` provider as `azapi` provider has no such top level attribute.

Note that we support for `resource_group_name` based on the fact that this is a well known attribute in `azurerm` to represent the residing resource group's name of the current resource. That's also the reason why we only support it at the top level, since there can be resources who has nested block, that also expose a `resource_group_name` and a `subscription_id` (e.g. `azurerm_data_share_dataset_blob_storage`), in which case the actual resource group can be different than the resource group in the current exported resource list (i.e. they have the same name, but from the different subscriptions).