`res` mode: Fix a bug that some resource types are not exported when `--include-role-assignment` is specified (#641)
* `res` mode: Fix a bug that some resource types are not exported when `--include-role-assignment` is specified

Previously, when `--include-role-assignment` is specified, the resource mode will list the specified resources together with the role assignments directly assigned on the resources via `azlist`, which in turn will query the specified resources via ARG's `Resource` table, using a predict like `id in (id1, id2,...)`.

The problem is that not all resource types are supported in the `Resource` table (e.g. template deployment resource). Also, it makes no sense to ask the user to specify the ARG table in the resource mode.

This PR updates the `azlist` dependency, use the new `ListByIds()` API to list those specified resources, together with the extension resources. This gets rid of querying from ARG to avoid un-supported resource types.

* ut